feat: Separate `Get-` cmdlet
- Add `Get-FactorioProfileSettings` cmdlet for retrieve the general module settings.
- Refactor `Get-FactorioProfile` to only retrieve profiles.
